requirements.txt
Flask==3.0.0
Pillow==10.1.0
geopy==2.4.1
